pyproject.toml
[project]
name = "mlops-club-rag"
version = "0.1.0"
description = "RAG-A-THON: Taking Python to Production + Cloud Engineering"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "arize-phoenix>=5.2.2",
    "llama-index-core>=0.11.17",
    "llama-index-vector-stores-pinecone>=0.2.1",
    "llama-index>=0.11.17",
    "openinference-instrumentation-llama-index>=3.0.0",
    "opentelemetry-exporter-otlp>=1.27.0",
    "opentelemetry-proto>=1.12.0",
    "opentelemetry-sdk>=1.27.0",
    "pinecone==5.3.1",
    "python-dotenv>=1.0.1",
    "reflex>=0.6.2.post1",
    "ruff>=0.6.9",
    "transformers>=4.45.2",
]
